+++ tests/test_accesslog.py
@@ -42,6 +42,7 @@ def test_post_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'client': 'test-client',
@@ -83,6 +84,7 @@ def test_post_array_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'method': 'POST',
@@ -129,6 +131,7 @@ def test_put_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'method': 'PUT',
@@ -166,6 +169,7 @@ def test_pdf_put_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'action': 'update',
@@ -214,6 +218,7 @@ def test_patch_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -257,6 +262,7 @@ def test_get_accesslog(app, model, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -301,6 +307,7 @@ def test_get_array_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -337,6 +344,7 @@ def test_pdf_get_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'format': 'json',
@@ -379,6 +387,7 @@ def test_get_prop_accesslog(app, model, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -420,6 +429,7 @@ def test_get_w_select_accesslog(app, model, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -460,6 +470,7 @@ def test_getall_accesslog(app, model, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -498,6 +509,7 @@ def test_getall_w_select_accesslog(app, model, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -545,6 +557,7 @@ def test_accesslog_file(model, postgresql, rc, request, tmpdir):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -654,6 +667,7 @@ def test_accesslog_file_stdin(
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'action': 'insert',
@@ -715,6 +729,7 @@ def test_accesslog_file_stderr(
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'action': 'insert',
@@ -757,6 +772,7 @@ def test_delete_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -796,6 +812,7 @@ def test_pdf_delete_accesslog(model, app, context):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -863,6 +880,7 @@ def test_pdf_ref_update_accesslog(model, app, context, tmpdir):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -913,6 +931,7 @@ def test_batch_write(model, app, context, tmpdir):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -957,6 +976,7 @@ def test_stream_write(model, app, context, tmpdir):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -1001,6 +1021,7 @@ def test_ns_read(model, app, context, tmpdir):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -1044,6 +1065,7 @@ def test_ns_read_csv(model, app, context, tmpdir):
     assert accesslog[-2:] == [
         {
             'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',